requirements.txt
Django
djangorestframework
django-rest-knox
django-environ
Pillow